Change event emitter logic and add favicon img file.

+++ contracts/Count.sol
@@ -17,16 +17,16 @@ pragma solidity ^0.5.6;
  */
 
 contract Count {
-  uint public count = 0;
-  address public lastParticipant;
+    uint256 public count = 0;
+    address public lastParticipant;
 
-  function plus() public {
-    count++;
-    lastParticipant = msg.sender;
-  }
+    function plus() public {
+        count++;
+        lastParticipant = msg.sender;
+    }
 
-  function minus() public {
-    count--;
-    lastParticipant = msg.sender;
-  }
+    function minus() public {
+        count--;
+        lastParticipant = msg.sender;
+    }
 }
